Rename chart_session to chart_section
