+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/observability/traces_data_source_profile/sub_profiles/otel.ts
@@ -18,7 +18,7 @@ import {
 import { DataSourceCategory, type DataSourceProfileProvider } from '../../../../profiles';
 import { extendProfileProvider } from '../../../extend_profile_provider';
 import { extractIndexPatternFrom } from '../../../extract_index_pattern_from';
-import { createChartSection } from '../accessors/chart_session';
+import { createChartSection } from '../accessors/chart_section';
 import { reContainsTracesApm, reContainsTracesOtel } from './reg_exps';
 
 export const createTracesOtelDataSourceProfileProvider = (
